fix: allow storage options provider without expires_at_millis (#5542)
Mainly for test use cases, it might be possible to return a credential
that never expires, and currently we just block the use case which is
unnecessary.

Also ensure that when user explicitly sets an AwsCredentialProvider in
rust, it is respected instead of getting only the initial credentials
and later being overwritten by the
DynamicStorageOptionsCredentialProvider.